Enable API cycling at runtime
Press T to cycle renderers/windows at runtime
Reenable D3D input callbacks
Fix quad vertex winding in CubePosCol.cpp
Enable gl face culling
Add MultiRenderer and MultiWindowWrapper, both of which are currently
unused

+++ src/ReadFile.cpp
@@ -16,7 +16,7 @@ std::vector<uint8_t> ReadFile(const std::string& filePath)
 	std::streampos fileLength = inFile.tellg();
 
 	std::vector<uint8_t> data;
-	data.resize(fileLength);
+	data.resize((size_t)fileLength);
 
 	inFile.seekg(0, std::ios::beg);
